@@ -1,5 +1,5 @@
 {
-  "name": "keycon",
+  "name": "name",
   "version": "0.0.0",
   "private": true,
   "dependencies": {
@@ -8,7 +8,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
     "react-scripts": "4.0.2",
-    "typescript": "^4.4.4"
+    "typescript": "^4.7.4"
   },
   "scripts": {
     "start": "react-scripts start",
