@@ -9,13 +9,13 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "typescript": "^4.3.2",
-    "ts-node": "^10.0.0"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   },
   "devDependencies": {
     "@types/cors": "^2.8.10",
     "@types/express": "^4.17.12",
-    "@types/express-session": "^1.17.4",
+    "@types/express-session": "^1.17.5",
     "@types/node": "^15.12.2",
     "@typescript-eslint/eslint-plugin": "^4.26.1",
     "@typescript-eslint/parser": "^4.26.1",
@@ -24,7 +24,7 @@
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-prettier": "^3.4.0",
-    "prettier": "^2.3.1",
-    "nodemon": "^2.0.7"
+    "nodemon": "^2.0.19",
+    "prettier": "^2.7.1"
   }
 }
