@@ -1,1 +1,5 @@
-export const port = parseInt(process.env.PORT || "3011", 10);
+export const PORT = parseInt(process.env.PORT || "3010", 10);
+
+export const log = (message: string): void => {
+  console.log(new Date().toLocaleString(), message);
+};
