@@ -1,19 +1,17 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-
 import { Provider } from "react-redux";
 
+import { store } from "./store";
 import App from "./App";
-import { store } from "./AppSlice";
-
 import "./index.css";
 
-const container = document.getElementById("root");
-const root = createRoot(container!);
-root.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>
-);
+const root = document.getElementById("root");
+if (root)
+  createRoot(root).render(
+    <React.StrictMode>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </React.StrictMode>
+  );
